requirements.txt
sentence-transformers
faiss-cpu
ollama
nltk
spacy
datasets
